+++ Recap_1_13/src/main/java/me/saurpuss/recap/commands/RecapCommand.java
@@ -7,35 +7,29 @@ import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
+import org.bukkit.command.TabCompleter;
 import org.bukkit.entity.Player;
+import org.bukkit.event.HandlerList;
+import org.bukkit.util.StringUtil;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 import java.util.logging.Level;
 
-/**
- * Command implementation for server side recap access.
- */
-public class RecapCommand implements CommandExecutor {
+public class RecapCommand implements CommandExecutor, TabCompleter {
 
-    /**
-     * Current plugin runtime.
-     */
     private Recap recap;
 
-    /**
-     * Command implementation constructor registered in Recap#onEnable()
-     *
-     * @param plugin dependency injection of the current plugin runtime
-     */
     public RecapCommand(Recap plugin) {
         recap = plugin;
     }
 
     /**
-     * Implementation of the /recap command:
-     * /recap           - Display the last 10 recaps
-     * /recap reload    - If the CommandSender has permission this reloads
-     *                    the plugin config.yml
-     * /recap [message] - Create an addition to the recap log
+     * /recap               - Display the last 10 recaps
+     * /recap reload        - If the CommandSender has permission this reloads
+     *                        the plugin config.yml
+     * /recap [message...]  - Create an addition to the recap log
      */
     @Override
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
@@ -51,24 +45,37 @@ public class RecapCommand implements CommandExecutor {
                 sender.sendMessage(ChatColor.RED + "You do not have permission!");
                 return false;
             }
+
+            // Reload config & manager
             recap.reloadConfig();
             recap.reloadRecapManager();
 
-            // TODO reload event registration
+            // Reload events if necessary
+            HandlerList.unregisterAll(recap);
+            recap.registerEvents();
 
-            for (Player player : Bukkit.getOnlinePlayers()) {
-                if (player.hasPermission("recap.notify")) {
-                    // TODO reload notify permission
-                    player.sendMessage(ChatColor.GREEN + "Reloaded Recap!");
-                }
-            }
-
+            // Notify relevant parties
+            for (Player player : Bukkit.getOnlinePlayers())
+                if (player.hasPermission("recap.reload.notify"))
+                    player.sendMessage(ChatColor.GREEN + "Reloaded plugin!");
             recap.getLogger().log(Level.INFO, "Finished reloading plugin!");
+
             return true;
         }
 
         // Save the arguments as a string to add to the recap log
         recap.getRecapManager().addRecap(sender, StringUtils.join(args, ' '));
         return true;
+    }
+
+    @Override
+    public List<String> onTabComplete(CommandSender sender, Command command, String alias, String[] args) {
+        if (!sender.hasPermission("recap.reload")) return null; // no permission, exit fast
+
+        // In case of additional commands use: Arrays.asList("reload", "other");
+        final List<String> COMMANDS = Collections.singletonList("reload");
+
+        return (args.length == 1) ? StringUtil.copyPartialMatches(args[0], COMMANDS,
+                new ArrayList<>()) : null;
     }
 }