+++ Recap_1_13/src/main/java/me/saurpuss/recap/events/JoinNotifyListener.java
@@ -6,22 +6,10 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
 
-/**
- * Event listener to allow players with the permission node recap.notify to get the last recaps
- * logged displayed in their chat. Only registered if 'show-on-join' in config.yml is true.
- */
 public class JoinNotifyListener implements Listener {
 
-    /**
-     * Runtime instance of the Recap plugin
-     */
     private Recap recap;
 
-    /**
-     * Constructor to get the event up and running
-     *
-     * @param plugin dependency injection of the current plugin runtime
-     */
     public JoinNotifyListener(Recap plugin) {
         recap = plugin;
     }
@@ -29,8 +17,6 @@ public class JoinNotifyListener implements Listener {
     /**
      * PlayerJoinEvent Listener that displays the last recap log to players with the recap.notify
      * permission.
-     *
-     * @param event on player join
      */
     @EventHandler
     public void onJoin(PlayerJoinEvent event) {
@@ -39,5 +25,4 @@ public class JoinNotifyListener implements Listener {
         Player player = event.getPlayer();
         recap.getRecapManager().getRecapLog().forEach(player::sendMessage);
     }
-
 }