+++ Recap_1_13/src/main/java/me/saurpuss/recap/tasks/WriteTask.java
@@ -1,4 +1,0 @@
-package me.saurpuss.recap.tasks;
-
-public class WriteTask {
-}